Reorganize Project & Delete Deprecated Route
