+++ src/app/api/ai-plugin/route.ts
@@ -1,5 +1,3 @@
-import { NextResponse } from "next/server";
-import { ACCOUNT_ID, PLUGIN_URL, COW_SUPPORTED_CHAINS } from "../../config";
 import {
   addressOrSymbolParam,
   addressParam,
@@ -9,6 +7,9 @@ import {
   SignRequestSchema,
   MetaTransactionSchema,
 } from "@bitte-ai/agent-sdk";
+import { NextResponse } from "next/server";
+
+import { ACCOUNT_ID, PLUGIN_URL, COW_SUPPORTED_CHAINS } from "@/src/app/config";
 
 export async function GET() {
   const pluginData = {
@@ -61,82 +62,6 @@ This assistant follows these specifications with zero deviation to ensure secure
       },
     },
     paths: {
-      // "/api/tools/cowswap": {
-      //   post: {
-      //     tags: ["cowswap"],
-      //     operationId: "swap",
-      //     summary:
-      //       "Posts unsigned order to CoW and returns Signable payload",
-      //     description:
-      //       "Given a partial order compute the minimum fee and a price estimate for the order. Return a full order that can be used directly for signing, and with an included signature, passed directly to the order creation endpoint.",
-      //     parameters: [
-      //       { $ref: "#/components/parameters/chainId" },
-      //       { $ref: "#/components/parameters/evmAddress" },
-      //       {
-      //         in: "query",
-      //         name: "sellToken",
-      //         required: true,
-      //         schema: {
-      //           type: "string",
-      //         },
-      //         description:
-      //           "The ERC-20 token symbol or address to be sold, if provided with the symbol do not try to infer the address.",
-      //       },
-      //       {
-      //         in: "query",
-      //         name: "buyToken",
-      //         required: true,
-      //         schema: {
-      //           type: "string",
-      //         },
-      //         description:
-      //           "The ERC-20 token symbol or address to be bought, if provided with the symbol do not try to infer the address..",
-      //       },
-      //       {
-      //         in: "query",
-      //         name: "receiver",
-      //         required: false,
-      //         schema: {
-      //           type: "string",
-      //         },
-      //         description:
-      //           "The address to receive the proceeds of the trade, instead of the sender's address.",
-      //       },
-      //       {
-      //         in: "query",
-      //         name: "sellAmountBeforeFee",
-      //         required: true,
-      //         schema: {
-      //           type: "string",
-      //         },
-      //         description:
-      //           "The amount of tokens to sell before fees, represented as a decimal string in token units. Not Atoms.",
-      //       },
-      //     ],
-      //     responses: {
-      //       "200": { $ref: "#/components/responses/SignRequestResponse200" },
-      //       "400": {
-      //         description: "Error quoting order.",
-      //         content: {
-      //           "application/json": {
-      //             schema: {
-      //               $ref: "#/components/schemas/PriceEstimationError",
-      //             },
-      //           },
-      //         },
-      //       },
-      //       "404": {
-      //         description: "No route was found for the specified order.",
-      //       },
-      //       "429": {
-      //         description: "Too many order quotes.",
-      //       },
-      //       "500": {
-      //         description: "Unexpected error quoting an order.",
-      //       },
-      //     },
-      //   },
-      // },
       "/api/tools/quote": {
         post: {
           tags: ["quote"],
