+++ src/app/api/tools/order/route.ts
@@ -1,7 +1,9 @@
 import { handleRequest } from "@bitte-ai/agent-sdk";
-import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { createOrder } from "./create";
+
+import { createOrder } from "@/src/lib/protocol/order";
+
+import type { NextRequest } from "next/server";
 
 export async function POST(req: NextRequest): Promise<NextResponse> {
   console.log("order/", req.url);
