+++ tests/util.spec.ts
@@ -1,33 +1,6 @@
-import { NextRequest } from "next/server";
-import { validateNextRequest } from "@/src/app/api/tools/util";
-import { zeroAddress } from "viem";
 import { OrderSigningUtils } from "@cowprotocol/cow-sdk";
 import { getTokenDetails, loadTokenMap } from "@bitte-ai/agent-sdk";
 import { COW_SUPPORTED_CHAINS } from "@/src/app/config";
-
-describe("validateRequest", () => {
-  it("should validate a real request", async () => {
-    const request = new NextRequest(
-      new Request("https://example.com", {
-        method: "POST",
-        headers: new Headers({
-          "mb-metadata": JSON.stringify({
-            accountId: "max-normal.near",
-            evmAddress: zeroAddress,
-          }),
-        }),
-        body: JSON.stringify({ test: "data" }),
-      }),
-    );
-
-    // Act
-    const result = await validateNextRequest(request);
-    // Assert
-    expect(result).toBeNull();
-  });
-
-  // domain: OrderSigningUtils.getDomain(parsedRequest.chainId),
-});
 
 describe("CoW Domain", () => {
   it("should get the domain", async () => {
